+++ modules/farm_rothamsted_quick/src/Plugin/QuickForm/QuickCombineHarvest.php
@@ -30,7 +30,7 @@ class QuickCombineHarvest extends QuickExperimentFormBase {
   /**
    * {@inheritdoc}
    */
-  protected $parentLogCategoryName = 'Harvest categories';
+  protected $parentLogCategoryName = 'Combine harvest categories';
 
   /**
    * {@inheritdoc}